+++ ultraviolet_cli/__init__.py
@@ -5,9 +5,10 @@
 # ultraviolet-cli is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
 
-"""Invenio module for custom Ultravaiolet commands"""
+"""Invenio module for custom Ultraviolet commands."""
 
 from .ext import ultravioletcli
+from .proxies import current_communities
 from .version import __version__
 
-__all__ = ('__version__', 'ultravioletcli')
+__all__ = ('__version__', 'ultravioletcli', 'current_communities')